+++ robotcode/language_server/robotframework/parts/debugging_utils.py
@@ -2,10 +2,19 @@ from __future__ import annotations
 
 import re
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, Generator, List, Literal, Optional, Union
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    AsyncGenerator,
+    Generator,
+    List,
+    Literal,
+    Optional,
+    Union,
+)
 
 from ....jsonrpc2.protocol import rpc_method
-from ....utils.async_itertools import async_dropwhile, async_takewhile
+from ....utils.async_itertools import async_dropwhile, async_next, async_takewhile
 from ....utils.async_tools import run_coroutine_in_thread
 from ....utils.logging import LoggingDescriptor
 from ...common.lsp_types import Model, Position, Range, TextDocumentIdentifier
@@ -14,6 +23,7 @@ from ..utils.ast import (
     Token,
     get_nodes_at_position,
     get_tokens_at_position,
+    range_from_node,
     range_from_token,
     tokenize_variables,
 )
@@ -89,8 +99,7 @@ class RobotDebuggingUtilsProtocolPart(RobotLanguageServerProtocolPart, ModelHelp
         re.UNICODE | re.VERBOSE,
     )
 
-    @staticmethod
-    def _get_all_variable_token_from_token(token: Token) -> Generator[Token, Any, Any]:
+    async def _get_all_variable_token_from_token(self, token: Token) -> AsyncGenerator[Token, Any]:
         from robot.api.parsing import Token as RobotToken
         from robot.variables.search import contains_variable
 
@@ -98,12 +107,13 @@ class RobotDebuggingUtilsProtocolPart(RobotLanguageServerProtocolPart, ModelHelp
 
             for sub_token in tokenize_variables(to, ignore_errors=ignore_errors):
                 if sub_token.type == RobotToken.VARIABLE:
-                    yield sub_token
-                    sub_text = sub_token.value[2:-1]
+                    base = sub_token.value[2:-1]
+                    if base and not (base[0] == "{" and base[-1] == "}"):
+                        yield sub_token
 
-                    if contains_variable(sub_text):
+                    if contains_variable(base):
                         for j in iter_all_variables_from_token(
-                            RobotToken(token.type, sub_text, to.lineno, to.col_offset + 2),
+                            RobotToken(token.type, base, to.lineno, to.col_offset + 2),
                             ignore_errors=ignore_errors,
                         ):
                             if j.type == RobotToken.VARIABLE:
@@ -114,7 +124,7 @@ class RobotDebuggingUtilsProtocolPart(RobotLanguageServerProtocolPart, ModelHelp
             match = RobotDebuggingUtilsProtocolPart._match_extended.match(name[2:-1])
             if match is not None:
                 base_name, _ = match.groups()
-                name = f"{name[0]}{{{base_name}}}"
+                name = f"{name[0]}{{{base_name.strip()}}}"
             yield RobotToken(e.type, name, e.lineno, e.col_offset)
 
     @rpc_method(name="robot/debugging/getEvaluatableExpression", param_type=EvaluatableExpressionParams)
@@ -133,20 +143,29 @@ class RobotDebuggingUtilsProtocolPart(RobotLanguageServerProtocolPart, ModelHelp
             if document is None:
                 return None
 
-            model = await self.parent.documents_cache.get_model(document)
+            namespace = await self.parent.documents_cache.get_namespace(document)
+            if namespace is None:
+                return None
 
-            node = (await get_nodes_at_position(model, position))[-1]
+            model = await self.parent.documents_cache.get_model(document)
+            if model is None:
+                return None
+
+            nodes = await get_nodes_at_position(model, position)
+            node = nodes[-1]
 
             if not isinstance(node, HasTokens):
                 return None
 
             token = get_tokens_at_position(node, position)[-1]
 
-            sub_token = next(
+            sub_token = await async_next(
                 (
                     t
-                    for t in RobotDebuggingUtilsProtocolPart._get_all_variable_token_from_token(token)
-                    if t.type == RobotToken.VARIABLE and position in range_from_token(t)
+                    async for t in self._get_all_variable_token_from_token(token)
+                    if t.type == RobotToken.VARIABLE
+                    and position in range_from_token(t)
+                    and await namespace.find_variable(t.value, nodes, position) is not None
                 ),
                 None,
             )
@@ -170,6 +189,7 @@ class RobotDebuggingUtilsProtocolPart(RobotLanguageServerProtocolPart, ModelHelp
     ) -> List[InlineValue]:
         async def run() -> List[InlineValue]:
             from robot.api import Token as RobotToken
+            from robot.parsing.model.blocks import Keyword, TestCase
 
             document = await self.parent.documents.get(text_document.uri)
             if document is None:
@@ -183,6 +203,15 @@ class RobotDebuggingUtilsProtocolPart(RobotLanguageServerProtocolPart, ModelHelp
 
             real_range = Range(view_port.start, min(view_port.end, context.stopped_location.end))
 
+            nodes = await get_nodes_at_position(
+                await self.parent.documents_cache.get_model(document), context.stopped_location.start
+            )
+
+            testcase_or_keyword_node = next((v for v in nodes if isinstance(v, (TestCase, Keyword))), None)
+            testcase_or_keyword_node_range = (
+                range_from_node(testcase_or_keyword_node) if testcase_or_keyword_node is not None else None
+            )
+
             result: List[InlineValue] = []
             async for token in async_takewhile(
                 lambda t: range_from_token(t).end.line <= real_range.end.line,
@@ -192,12 +221,22 @@ class RobotDebuggingUtilsProtocolPart(RobotLanguageServerProtocolPart, ModelHelp
                 ),
             ):
                 added: List[str] = []
-                for t in RobotDebuggingUtilsProtocolPart._get_all_variable_token_from_token(token):
+                async for t in self._get_all_variable_token_from_token(token):
                     if t.type == RobotToken.VARIABLE:
-                        upper_value = t.value.upper()
-                        if upper_value not in added:
-                            result.append(InlineValueEvaluatableExpression(range_from_token(t), t.value))
-                            added.append(upper_value)
+                        position = range_from_token(token).start
+
+                        var = await namespace.find_variable(
+                            t.value,
+                            nodes
+                            if testcase_or_keyword_node_range is not None and position in testcase_or_keyword_node_range
+                            else None,
+                            position,
+                        )
+                        if var is not None:
+                            upper_value = t.value.upper()
+                            if upper_value not in added:
+                                result.append(InlineValueEvaluatableExpression(range_from_token(t), t.value))
+                                added.append(upper_value)
 
             return result